+++ tarantool/src/fiber/async.rs
@@ -36,6 +36,10 @@ pub mod watch;
 
 pub use mutex::Mutex;
 
+#[cfg(feature = "async-std")]
+pub use async_std;
+pub use futures;
+
 /// Error that happens on the receiver side of the channel.
 #[derive(thiserror::Error, Debug, PartialEq, Eq)]
 #[error("sender dropped")]